Split filesytem deployment and image partitioning
The image-setup action ended up being really two things, doing all the
image partitioning and creating the filesystems as well as deploying the
currently setup filesystem onto it and create its fstab and kernel
cmdline. This especially showed when what was deployed onto the images
wasn't the filesystem but an ostree commit as the latter doesn't have
e.g. an /etc on /

Split the action into two new actions:
* image-partition, which does all the partitioning and filesystem creation
  (and mounting those)
* filesystem-deploy, which deploys the current filesystem onto the image
  and sets up fstab + /etc/kernel/cmdline

For fstab setup and kernel commandline the required information is
stashed into the general context such that both the ostree deploy action
and the filesystem deploy action can re-use it for their particular
setups

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>