Improve logging
Some action can take a good amount of time (packing, unpacking) but gave
no indication that they were running. Improve this by logging whenever
an action starts (using the description) so it's easier to determine
what's going on. Also log at the end (mostly so there is a timestamp)
and in case fakemachine isn't supported, which is a less well tested
usecase.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>