Add origins mechanism usage for RawAction
Switch 'RawAction' to common convention, so use `origin` and `source` for indicating
the file to use instead of `source` and `path`, respectively.

Now it is possible to use any available origin.

Recipe example:
```
 - action: raw
    description: Flash u-boot
    origin: filesystem
    offset: {{ sector 768 }}
    source: /usr/lib/u-boot/u-boot.img
```

Backward compatibility is supported for Raw action.
Allow to use old deprecated syntax based on 'source' and 'path'
properties. Warn user if old syntax is used.

Signed-off-by: Denis Pynkin <denis.pynkin@collabora.com>