Bind mount disk devices in chroot
Some operations need access to the disk device backing the image that's
being build, in particular for installing grub-pc. e.g. grub-install needs to write to the MBR to add
it's boot imgae and probed information.

To allow this bind mount the image device file,  all devices files for
the created partitions and the /dev/disk directory (for udev alias symlinking)
into a chroot when running commands in it.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>