Fix emptyDir
There was a mistake in os.RemoveAll() usage, only the basename of the
file was passed, this can't work. The error was silently ignored.

This error is easily spotten if you use OstreeCommit, this function is
supposed to make `/dev` empty, but in the resulting ostree the `/dev`
directory is not empty.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>