Add format-image action
format-image creates an image file and formats it with a filesystem. It
is meant to be followed by the 'filesystem-deploy' action.

This action is useful to prepare images that can be copied "as is" on a
block device, for example one could prepare a rootfs image and install
it with a command such as:

    dd if=rootfs.img of=/dev/sda1

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>