debootstrap_action: add defaults for Mirror and Components
debos creates a sources.list file based on Mirror and Components from
the specification. However the only mandatory property for the
debootstrap action is suite.

Currently not specifying Mirror and/or Components results in a faulty
sources.list. This patch adds default values for the Mirror and
Components properties fixing the issue. It also document the defaults.

Signed-off-by: Peter Senna Tschudin <peter.senna@collabora.com>