Add support for HFS/HFS+ filesystems in image-partition action
HFS and HFS+ filesystems are not handled the same by parted and the
kernel: for instance, HFS+ is named 'hfsplus' in the kernel, but
parted expects 'hfs+' or 'hfsx' (case-sensitive variant).

This patch fixes debos behavior in that case and implements the
proper command-line options when formatting these filesystems.

debos will now recognize and properly handle the 'hfs', 'hfsplus'
and 'hfsx' filesystem types.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>