Add warning on unmount failure
We don't really have a way atm to fail based on cleanup failures as
they're done in deferreds, so that would a bit need a bit more
refactoring. For now just add awarning message if unmounts failed. While
this isn't actually typically harmful for runs that can use fakemachine
(as the postprocessing happens in the outer debos), for runs without
fakemachine it can cause the postpressing to work on an image file
that's still dirty.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>