actions: ostree-deploy: Add explicit GC at end of run
Various Ostree object in particular OstreeRepo keeps file descriptors
open until the C Gobject is finalized. While this is not a big problem
when doing C where the dropping of refs is rather explicit anyway, it's
problematic in GC'd language where the refs might not be dropped for a
long time.

In particular for debos this is problematic when the clean up of an
image-partition action is trying to unmount its various partitions,
while the still alive objects from an ostree-deploy haven't been GC'd
just yet.

This really should be fixed by going over the ostree Go bindings and add
explicit deinitializers rather then relying on finalizers for cleanup.
However until that time explicitly running the GC at the end of the Run
function should nicely paper over the issue.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>