actions: Add recipe action
This action includes the recipe actions at the given path, and can
optionally overrides or adds new template variables.

To allow recipes to operate both stand-alone and when included, template
variable "included_recipe" is added and set to "true" to indicate that
the recipe is part of a longer chain.
This allows a recipe to skip the unpack step if it's not the first to be
executed or to optionally skip a pack action when it's not the last.

For compatibility, both the parent recipe and all included recipes have
to be for the same architecture.

Limitations of combined recipes are equivalent to limitations within a
single recipe (e.g. there can only be one image partition action).

Print-recipe option is saved in DebosContext to allow each "recipe action"
to print its source during Parse.

Signed-off-by: Frédéric Danis <frederic.danis@collabora.com>