Revert "Fix debootstrap failed with multiple components"
This reverts commit d6e16be95b2def8eb47aa9e79e9a608d0eb5d905.

Not passing along components to the debootstrap second stage makes the
recipe fail iff the components other then those of debian are uses (e.g.
not main!).

Revert this patch as it makes it impossible to build images for
distributions not having a main component, while the problem with
multiple components can easily be worked around in a recipe

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>