actions: Dump actions for each recipe
Dump Recipe structure in Recipe.Parse() method to allow to do it before
Verify() method is called for a given recipe or sub-recipe.

Sub-recipes are loaded by recipe action's Verify(). Dump of Recipe
structure can be done in Verify() or Parse() methods.
Doing it in Recipe.Parse() allows to have it done in one place.

Signed-off-by: Frédéric Danis <frederic.danis@collabora.com>