Split DebosContext in generic and recipe specific parts
To be able to load recipe from directory not under main recipe
directory, new "recipe" action needs to be able overwrite some
properties, architecture and recipe path. But other properties
needs to be shared between actions, as some of them can change
their values.
To allow this, move shared properties in new CommonContext structure
and add to remaining properties a new pointer to the CommonContext.

Signed-off-by: Frédéric Danis <frederic.danis@collabora.com>