Drop the included_recipe template var
When adding support for including recipes to debos a special template
variable (`included_recipe`) was added such that a recipe could
determine whether it was the "main" recipe or an included recipe. The
thinking was that this would allow recipes to operate in both stand-alone
and included mode easily by potentially making it only execute some
actions depending on the mode.

In hindsight this is was too simplistic. For example for recipes which
operate on a pre-prepared rootfs one may want to skip unpacking such a
rootfs tarball as a *first* step iff the "main" recipe prepared it
already (e.g.  by having a debootstrap action). However the
`included_recipe` variable can't really communicate that information, so
you end up with a recipe that (if included) must either assume the
preperation has been done or always do it itself defeating the point
somewhat.

A better solution for these cases is to have such a recipe accept a
specific template variable to disable prepearation actions and have the
"main" recipe set that explicitly as needed depending on the situation.

Given that drop the included_recipe template variable at least for now
(before it starts seeing usage), something like it can potentially be
re-added in the future when there is a more clear and meaningful
use-case for it.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>