Move target bind mount for scripts to tmp
`/script` directory is left in target rootfs. Moving the target bind
mount to `/tmp` allows to avoid additional step for removing this directory
for all projects calling own scripts in chrooted environment.
This commit change the behavior only for "chroot" method with `systemd-nspawn`,
which explicitly mount tmpfs to /tmp for container. Rationale:
https://github.com/systemd/systemd/commit/bbb99c30d01a8bcdc27fb151cc6376a7877a6b07

Fixes #151

Signed-off-by: Denis Pynkin <denis.pynkin@collabora.com>