pack,unpack: Save and restore xattrs and file capabilities in tarballs
Add options to `tar` to tell it to save and restore extended attributes:

* `--xattr`: enable extended attributes support
* `--xattrs-include=*.*`: tell `tar` to include every extended attribute
  since by default `tar` only stores attributes in the `user.*` namespace

This fixes the `pack` and `unpack` actions when dealing with tools like
`ping` which on modern distributions have been switched to use file
capabilities to do privileged operations without being setuid, as they are
based on extended attributes.

This relies on `tar` being GNU Tar >= 1.27, released in 2013 and shipped since
Debian Jessie.

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>