Add pacman action
Add a pacman action, whereby pacman is the Arch package manager akin to
apt for Debian. This allows us to install packages, after the initial
pacstrap action (or equivalent).

Akin to the pacstrap action - there is a Debian package for pacman. Plus
upstream pacman continuously builds and tests pacman against Arch,
Fedora and Debian.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>

v2: [Ed]
 - Drop --color never as it is implicit
 - Drop --noprogressbar likewise
 - Reorder remaining arguments
 - Style tweaks

v3: [Emil]
 - Run the lot through gofmt -s -w
 - Commit message
 - Sort the action alphabetically

v4: [Emil]
 - Drop quotes around pacman

Signed-off-by: Ed Smith <ed.smith@collabora.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
---

Similar to the pacstrap action, the formatting is a bit weird props to
gofmt. Happy to drop that hunk.