pacstrap: Import all keyrings in /usr/share/pacman/keyrings
With this change, pacman-key will populate the pacman keyring in the
chroot with EVERY keyring found in /usr/share/pacman/keyrings on the
host.

For example, assuming you're building a Manjaro image, you need to
populate the chroot with the two keyrings 'archlinux' and 'manjaro'.
Assuming that both keyrings are installed on the host, then you have two
ways to achieve that
- run `pacman-key --populate archlinux manjaro`, which install exactly
  these two keyrings in the chroot.
- run `pacman-key --populate`, which installs every keyring found in
  /usr/share/pacman/keyrings.

This commit makes the assumption that the host has already all the
keyrings needed in /usr/share/pacman/keyrings, and that it's desirable
to install them all in the chroot.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>