Reverse the exitcode initialization value
Fakemachine is subject to panic and causes Debos to exit success due to
the current logic of the exitcode. For example, the fakemachine function
CopyFileTo() panics if the file is missing.

In the case of a panic, the function never returns. Thus, the exitcode
cannot be set to 1 and Debos exits with 0.

This commit reverses the logic of the exitcode: it is initialized to 1
(i.e. failure), and it is set to 0 (i.e. success) only if Debos has
reached the end of the things it has to do (or for the help message).

Co-authored-by: Gaël PORTAY <gael.portay@collabora.com>
Co-authored-by: Santosh Mahto <santosh.mahto@collabora.com>

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>