action: run: expose context.ImageMntDir for recipes which do not call filesystem-deploy
When a command is ran using `action: run` some environment variables are exposed to the script.

`$ROOTDIR` is set to `context.Rootdir=/scratch/root` which is a temporary
directory meant to hold the rootfs until `action: filesystem-deploy` is ran, which
copies from `context.Rootdir` to `context.ImageMntDir`. After the filesystem is
deployed, `context.Rootdir` is set to `context.ImageMntDir`.

The mount points defined under action: image-partition are mounted under
`context.ImageMntDir`, so if the filesystem is never deployed, there is no
(simple) way for the script to know where the images are mounted to.

So, expose `context.ImageMntDir` to the script as `$IMAGEMNTDIR` so the recipe
can know where the image is mounted.

Resolves: #204

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>