actions/image_partition: sort mountpoints by position in filesystem hierarchy before mounting
Currently the `mountpoints` are mounted in the fakemachine in the order defined
by the recipe. This is confusing for users so sort the mountpoints in a sane
order before they are mounted.

This patch sorts the mountpoints so that (if defined) the root partition would
always be mounted first, then in ascending order of how many slashes ('/')
are in the mountpoint so that the parents are mounted first.

The current verbatim syntax for mountpoints where the recursive parent filesystems
are mounted before the children could be as follows:
```
  - action: image-partition
    imagename: mountpoints-order.img
    imagesize: 5MB
    partitiontype: gpt
    partitions:
      - name: root
        fs: ext4
        start: 0M
        end: 1M
      - name: opt
        fs: ext4
        start: 2M
        end: 3M
      - name: opt_test
        fs: ext4
        start: 4M
        end: 5M
    mountpoints:
      - mountpoint: /
        partition: rootfs
      - mountpoint: /opt
        partition: opt
      - mountpoint: /opt/test
        partition: opt_test
```

After mounting the fakemachine shows:
```
$ mount
...
/dev/vda1 on /scratch/mnt type ext4 (rw,relatime)
/dev/vda2 on /scratch/mnt/opt type ext4 (rw,relatime)
/dev/vda3 on /scratch/mnt/opt/test type ext4 (rw,relatime)

$ df -h
...
/dev/vda1       940K   22K  851K   3% /scratch/mnt
/dev/vda2       955K   22K  866K   3% /scratch/mnt/opt
/dev/vda3       940K   21K  852K   3% /scratch/mnt/opt/test
```
which is the correct order.

If the user defines the recipe to mount these out of order, for example:
```
- mountpoint: /opt
  partition: opt
- mountpoint: /opt/test
  partition: opt_test
- mountpoint: /
  partition: root
```
Without this patch, no error is shown but the system has mounted the partitions
out of order (which is confusing for the user):
```
$ mount
...
/dev/vda2 on /scratch/mnt/opt type ext4 (rw,relatime)
/dev/vda3 on /scratch/mnt/opt/test type ext4 (rw,relatime)
/dev/vda1 on /scratch/mnt type ext4 (rw,relatime)

$ df -h
...
/dev/vda1       940K   21K  852K   3% /scratch/mnt
```

With this patch, the recipe would show:
```
$ mount
...
/dev/vda1 on /scratch/mnt type ext4 (rw,relatime)
/dev/vda2 on /scratch/mnt/opt type ext4 (rw,relatime)
/dev/vda3 on /scratch/mnt/opt/test type ext4 (rw,relatime)

$ df -h
...
/dev/vda1       940K   22K  851K   3% /scratch/mnt
/dev/vda2       955K   22K  866K   3% /scratch/mnt/opt
/dev/vda3       940K   21K  852K   3% /scratch/mnt/opt/test
```

Resolves: #177

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>