actions/image_partition: trigger udev rules after image-partition completion
The image is locked while partitioning and mounting, which has a side effect of
not running udev rules so there are no entries for the freshly partitioned disk
under the children of `/dev/disk`.

Currently the workaround is to call partprobe in the recipe, which creates the
entries but we should do this as part of the image-partition action so that no
magic is needed in the recipe.

Fixes: f91e2b2c2789 ("Fix partition races while partitioning")
Resolves: #154

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>