actions/image-partition: add support for setting GPT partition label
Debos currently sets the GPT partition label to the `Name` property of the
Partition and does not allow the GPT partition label to be set from the recipe.

In some situations setting the GPT partition label is required, so add the
property `PartLabel` to the Partition and set the GPT partition label to the value
of that property. By default, the `PartLabel` property is set to the `Name`
property of the Partition to not break backwards compatibility.

Resolves: #202

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>