actions/run: Improve run action verification
For the run action, debos does not complain even if both "script" and
"command" are empty which can occur due to a typo or accidentally
missing the tag.

As the documentation already mentions that one of "script" or
"command" is mandatory for the run action, let's detect the invalid
scenario and complain to the user instead of silently continuing.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>