action/debootstrap: allow to use files outside of RecipeDir
Added stages "Verify" and "PreMachine" allowing to use secret files
outside of recipe directory tree (both absolute and relative paths
are supported).
Verification stage ensures that configuration files from options
are existing on file system.

This is needed for cases if secret files are not the part of Debos
recipe directory and useful for CI.

Signed-off-by: Denis Pynkin <denis.pynkin@collabora.com>