actions/image-partition: Consistently use artifactdir
The "image-partition" action uses artifactdir in the
PostMachineCleanup() step but not during the PreMachine() and
PreNoMachine() steps. Also, other debos actions such as pack, unpack
are relative to the specified "artifactdir" when invoking debos.

Fix this inconsistency by bringing image-partition action to the
fold. Also update the documentation to clarify this.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>