Docker tests should run without fakemachine
Since we have the support for the new user-mode-linux backend in the
docker container, Debos defaults to attempting to use that.
Unfortunately that isn't possible to run inside the Docker Hub where the
test suite runs, so let's explicitly disable the fakemachine library
when running the Docker Hub test suite.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>