image-partition: use fsuuid to set the filesystem UUID
This patch adds functionality that allows you to specify
the UUID of your file system instead of getting a random one.
Some use cases include reproducible builds and software updates.

Currently, only btrfs, xfs, ext2, ext3, and ext4 are supported.

The uuid library is used for verifying the input value of
the FSUUID property (specified by the user on the
image-partition action). The verification is based on RFC 4122
and DCE 1.1: Authentication and Security Services. Check
https://github.com/google/uuid for more details.

Signed-off-by: Nguyen Thi Huong <huong4.nguyenthi@toshiba.co.jp>
Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>