actions/image-partition: truncate filesystem label to maximum supported length
Added optional fslabel property to partition to allow truncation of filesystem
label and allow user to modify filesystem label without modifying the name.
Filesystem label defaults to the name property of the partition.

The filesystem label can be up to 11 characters long for {v}fat{12|16|32},
16 characters long for ext2/3/4, 255 characters long for btrfs, 512 characters
long for hfs/hfsplus and 12 characters long for xfs.

Fixes: #251

Suggested-by: Christopher Obbard <chris.obbard@collabora.com>
Signed-off-by: Vignesh Raman <vignesh.raman@collabora.com>
Signed-off-by: Dylan Aïssi <dylan.aissi@collabora.com>