Separate out publishing to DockerHub
This allows it to be independently disabled by a check to see
whether its credentials are likely to be valid, and prevents
forked repositories from having failures in this step.

We're forced to check every step, because both secrets. and
env. are only available at the point of steps, not at the level
of jobs (so simply disabling the entire DockerHub job based on
either of those value sources is not possible).