command: Don't log special characters from cmd output
Currently we print to the log to stdout whatever the downstream
command prints to stdout|stderr and let the shell on the host deal
with the handling of any special characters. In the case of some
non-interactive shells (like GitLab CI) any lines printed to its
stdout with special characters (like \b or \r) are simply not
printed to the web interface.

Some commands used by actions (apt, mkfs...) use these special
characters to print progress to the user, which is useful in
interactive environments but can completely break non-interactive
usage like in Debos.

Let's fix these environments by removing the offending characters
before printing the output to the shell.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>