Isolate from host keyrings
This provides much finer grained control of the trust base
of the generated rootfs, and the keyrings present on it. The
default is to use the system default (_not_ installed!) keyrings,
as pacman-key itself would

This change necessitates the use of the new facility in Command
to run a process with private bind mounts, since pacman-key is
hard-coded to use /usr/share/pacman/keyrings for all keyrings
that can be considered in the populate step. The alternative is
to replicate the logic of pacman-key itself.

Also, remove automatic copy of public keyring which is now obsolete.