Add a new CheckEnv stage for checking command availability
Verify (which should be called Prepare) must run before the
fakemachine is created, in order to setup the arguments for the
fakemachine itself. It then has to run again in the fakemachine
in order to rebuild any necessary internal state.

Command checking can only occur in the same environment as run, as
commands that are in the path when running may not be available to
the regular user (e.g. Debian installs debootstrap in sbin).