Add a separate chmnt program
chmnt is a small utility which creates private bind mounts for
a target program. It enables debos to conveniently permit both
bind mounts and an altered environment for non-chroot commands.

One reason you might want to do this is to enable hard-coded
paths in applications that debos needs to call to be altered,
even if those paths are within /usr (which is shared between
the fakemachine and the host, and is therefore alterable on
neither directly).