actions/image-partition: allow extended and logical
This commit allows us to create extended or logical
partitions on MBR partition table (msdos) in order to
address the following issues:

* Because part-type for parted is hardcoded with primary,
we cannot create more than 4 partitions using extended and
logical.

* Even if we manage to create extended and logical partitions,
the verificaton step fails because the Verify function assumes
that the partition number is always sequetially incremental.

This commit makes it possible that if the number of partitions
is more than 4, 3 of them are automatically created as primary,
1 of them is extended, and the remaining partitions are logical.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>
Signed-off-by: Daichi Fukui <daichi1.fukui@toshiba.co.jp>