actions/recipe: Only run cleanup function if needed
When running nested recipes using the recipe action,
if an action in the child recipe exits early (e.g. on
failure) the remaining actions do not run but the
cleanup functions for these remaining actions still
run even though there is nothing to cleanup.

This doesn't follow how Debos handles running standalone
recipes, so modify the recipe action to only run action
cleanup functions when the associated action needs to be
cleaned up.

Fixes: 74df488fb6b0 ("actions: Add recipe action")
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>