actions: overlay: Check source exists in recipedir early
If an overlay (with a source from the recipe directory which is unlikely
to change during execution) doesn't exist, the recipe errors out during
execution.

Change the logic to check early (in the verification stage) if the
overlay exists on the host filesystem inside the recipe directory, if
the overlay source is from the recipedir.

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>