actions: image-partition: Remove extra warning about removing loop device
When removing loop devices, we retry after a delay with
a timeout. Each time the loop device removal fails, a
warning message is printed to the log:

  "Loop dev couldn't remove could not remove, device in use, waiting"

This is pointless and could confuse the user. An error
while performing the loop device removal & retry is
logged seperately so let's remove the duplicate (and
confusing) log message.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>