Add helper to retrieve origin
The origins hash table is inside the commonContext structure which is
shared by all Contexts. This mostly works correctly as most origins are
not specific to the recipe that's executed but to either the run or the
progress of a run. Such as "filesystem" or "artifacts".

However this is not the case for the "recipe" origin. To resolve that
add a helper function which returns the current context RecipeDir and
otherwise simply the respective value in the common Origins table.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>