Use shared cache for multiple jobs
Every CI runs mostly builds the same base docker image, but as a
per-commit cache was used it was done from scratch every time. To
improve that switch to a cache key that's not specific to a commit. Also
always pull the base docker containers, which will trigger a rebuild
whenever those are updated to keep things fresh.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>