ci: Run tests inside test containers
For faster developer feedback, kick off a test build and the
unit test suite inside pre-made containers. This allows the
job to fail-fast since a developer doesn't need to wait for
a Docker container to be built.

This also has the benefit of testing Debos on various
platforms other than the final Docker container.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>