docs: list possible values for origin
Getting the raw action to work with an actual file was somewhat
confusing for me.

Looking at the code, the origin is searched for in context.Origins,
which is only set in three places:

 $ git grep Origins
 download_action.go:     context.Origins[d.Name] = originPath
 filesystem_deploy_action.go:    context.Origins["filesystem"] = context.ImageMntDir
 ostree_deploy_action.go:                context.Origins["filesystem"] = context.ImageMntDir

In addition, "recipe" is special-cased in DebosContext.Origin.

List all four options to hopefully make it less confusing for the
next person.