godoc: fix mangled yaml code blocks, take two
Unfortunatley, 191c474477af4cc679779dd4eb70f28c1d9dbd62 was not
enough to fix the yaml code block rendering on godoc.org.

The problem is that a block that starts with "-", even if indented,
is interpreted as a list.

The only solution I found was to pull the "Yaml syntax:" line
into the code block as a comment.

This fixes the rendering of the code block as verified by running
"go doc actions" on the command line (see below):

Rendering of the YAML block using "go doc actions | head -20",
before this change:

	package actions // import "github.com/go-debos/debos/actions"

	Package 'actions' implements 'debos' modules used for OS creation.

	# Apt Action

	Install packages and their dependencies to the target rootfs with 'apt'.

	Yaml syntax:

	  - action: apt recommends: bool unauthenticated: bool update: bool packages:
	  - package1
	  - package2

	Mandatory properties:

	- packages -- list of packages to install

	Optional properties:

After this change:

	package actions // import "github.com/go-debos/debos/actions"

	Package 'actions' implements 'debos' modules used for OS creation.

	# Apt Action

	Install packages and their dependencies to the target rootfs with 'apt'.

	    # Yaml syntax:
	    - action: apt
	      recommends: bool
	      unauthenticated: bool
	      update: bool
	      packages:
	        - package1
	        - package2

	Mandatory properties:

	- packages -- list of packages to install