tests: process json files throgh jq prior to comparing
There are various whitespace changes that can occur in json files. Use
the same pretty-format in both files to ensure the output is stable and
consistent.

Note: we need to explicitly call "bash -c" here otherwise, we'll end up
with pure POSIX shell (sh), where anonymous pipes are not a thing.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>